Change claim id to long from int for AB#16876.

+++ Apps/Encounter/src/Models/Claim.cs
@@ -27,7 +27,7 @@ namespace HealthGateway.Encounter.Models
         /// Gets or sets the ClaimId.
         /// </summary>
         [JsonPropertyName("claimId")]
-        public int ClaimId { get; set; }
+        public long ClaimId { get; set; }
 
         /// <summary>
         /// Gets or sets the Service Date.
